backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.14.0
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
jwt==1.4.0
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
